Bump to 0.0.7 alpha
[MD5 hash]
Off the bat I noticed I hadn't properly updated the webMD5 files web link.
   That's was a big issue caused by a small mistake. :(
I've also updated the hash's name and removed the old one.

Script Hash check:
I've moved this chunk of code to the bottom of the script which enables things to be kept safe still, but also allows for the default 'header' and 'usage' test to be printed if there's a hash mismatch and no flags were given.

Added the -s flag to the usage output.

Adding dates to backup folders.